Function that prints all the elements of a list_t list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -10,7 +10,7 @@
 */
 size_t print_list(const list_t *h)
 {
-	int s = 0;
+	size_t s = 0;
 
 	while (h)
 	{
